_job/generate/requirements.txt
functions-framework
google_cloud_bigquery==3.13.0
requests==2.31.0
pyyaml==6.0.1